[ecmascript6]Add class for media browser.
