+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules
+js/build.js
+js/build.min.js
