+++ README.md
@@ -2,3 +2,8 @@
 Javascript library to capture or pick  images and or audio from your PC
 
 Well...just images at the moment :)
+
+ ## Run Demo
+ * npm install
+ * npm run build
+ * npm start
