+++ index.html
@@ -1,0 +1,25 @@
+<!doctype html>
+<html>
+<head>
+	<title>Media Picker</title>
+</head>
+<body>
+	<img id="photo" style="width:320px;height:240px;">
+	<h2>A HTML page...</h2>
+
+	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut eget dictum tortor, eget lacinia diam. Mauris consectetur dapibus gravida. Vivamus lobortis dapibus augue condimentum interdum. In et est id sem commodo dictum. Cras quis enim tincidunt, aliquam
+		orci convallis, tincidunt dui. Aenean libero neque, pulvinar et ex sed, dapibus vehicula ligula. Sed nulla risus, molestie in eros non, sagittis porta risus. Quisque cursus ex porta, tincidunt lacus id, blandit eros. Mauris ac justo sollicitudin, aliquam
+		enim quis, fringilla nisl. Aenean vulputate eget orci a imperdiet. Fusce ultrices orci sed sollicitudin dapibus. Suspendisse accumsan ac diam at fringilla.</p>
+	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque viverra mi elit, a interdum velit ullamcorper quis. Donec ut justo accumsan, euismod dolor eget, gravida libero. Praesent accumsan massa vitae tortor facilisis, vel efficitur quam
+		lobortis. Etiam sapien erat, imperdiet non porta at, auctor quis ex. Phasellus quis libero eget massa scelerisque fringilla vel mollis mauris. Nullam at volutpat odio, sit amet faucibus dui. Ut lacinia sodales nibh, ut sollicitudin lorem dictum sed.
+		Mauris tristique ligula velit, euismod auctor quam mattis sed. Etiam non leo et nisi porttitor sollicitudin. Donec dictum urna et elit facilisis consequat ultrices a felis. Aenean arcu massa, pellentesque in augue ac, condimentum tempus elit. Fusce
+		laoreet mi a eros faucibus, ac posuere nisi commodo. Sed placerat diam sit amet finibus pretium.</p>
+	<p>Maecenas vitae mi lorem. Praesent rutrum enim justo. Proin tempus velit non lacus interdum, ac sagittis justo malesuada. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id eros vitae
+		sem iaculis commodo id ac nisl. Vestibulum id quam diam.</p>
+
+	<button id="capturar">Capture from webcam</button>
+	<button id="galeria">Browse from my PC</button>
+	<script src="js/build.min.js"></script>
+</body>
+
+</html>
