--- media-picker-demo.html
@@ -1,43 +1,0 @@
-<!doctype html>
-
-<html>
-	<head>	
-		<script type="text/javascript" src="media-picker-helper.js"></script>
-		<script type="text/javascript">
-			window.onload = function() {
-				document.getElementById('capturar').addEventListener('click', function() {
-					MediaBrowserHelper.capture()
-						.then(function(imageDataUrl) {
-							document.getElementById('photo').src = imageDataUrl;
-						}).catch(function(e) {
-							document.getElementById('photo').src = '';
-							alert('Error tomando foto: ' + JSON.stringify(e));
-						});
-				});
-				document.getElementById('galeria').addEventListener('click', function() {
-					MediaBrowserHelper.gallery()
-						.then(function(imageDataUrl) {
-							document.getElementById('photo').src = imageDataUrl;
-						}).catch(function(e) {
-							document.getElementById('photo').src = '';
-							alert('Error usando foto local: ' + JSON.stringify(e));
-						});
-				});
-			};
-		</script>
-	</head>
-	<body>
-		<div>
-			<img id="photo" style="width:320px;height:240px;">
-		<div>
-
-		<h2>A HTML page...</h2>
-
-		<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut eget dictum tortor, eget lacinia diam. Mauris consectetur dapibus gravida. Vivamus lobortis dapibus augue condimentum interdum. In et est id sem commodo dictum. Cras quis enim tincidunt, aliquam orci convallis, tincidunt dui. Aenean libero neque, pulvinar et ex sed, dapibus vehicula ligula. Sed nulla risus, molestie in eros non, sagittis porta risus. Quisque cursus ex porta, tincidunt lacus id, blandit eros. Mauris ac justo sollicitudin, aliquam enim quis, fringilla nisl. Aenean vulputate eget orci a imperdiet. Fusce ultrices orci sed sollicitudin dapibus. Suspendisse accumsan ac diam at fringilla.</p>
-		<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque viverra mi elit, a interdum velit ullamcorper quis. Donec ut justo accumsan, euismod dolor eget, gravida libero. Praesent accumsan massa vitae tortor facilisis, vel efficitur quam lobortis. Etiam sapien erat, imperdiet non porta at, auctor quis ex. Phasellus quis libero eget massa scelerisque fringilla vel mollis mauris. Nullam at volutpat odio, sit amet faucibus dui. Ut lacinia sodales nibh, ut sollicitudin lorem dictum sed. Mauris tristique ligula velit, euismod auctor quam mattis sed. Etiam non leo et nisi porttitor sollicitudin. Donec dictum urna et elit facilisis consequat ultrices a felis. Aenean arcu massa, pellentesque in augue ac, condimentum tempus elit. Fusce laoreet mi a eros faucibus, ac posuere nisi commodo. Sed placerat diam sit amet finibus pretium.</p>
-		<p>Maecenas vitae mi lorem. Praesent rutrum enim justo. Proin tempus velit non lacus interdum, ac sagittis justo malesuada. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id eros vitae sem iaculis commodo id ac nisl. Vestibulum id quam diam.</p>
-
-		<button id="capturar">Capture from webcam</button>
-		<button id="galeria">Browse from my PC</button>
-	</body>
-</html>
