+++ npm-debug.log
@@ -1,0 +1,40 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/bin/nodejs', '/usr/bin/npm', 'run', 'build' ]
+2 info using npm@2.14.20
+3 info using node@v4.4.0
+4 verbose run-script [ 'prebuild', 'build', 'postbuild' ]
+5 info prebuild media-picker@1.0.0
+6 info build media-picker@1.0.0
+7 verbose unsafe-perm in lifecycle true
+8 info media-picker@1.0.0 Failed to exec build script
+9 verbose stack Error: media-picker@1.0.0 build: `browserify js/app.js > js/build.js -t [babelify --presets [es2015 react] ] && uglify -s js/build.js -o js/build.min.js && && watchify js/app.js -o js/build.min.js -v -t [ babelify --presets [ react ] ]`
+9 verbose stack Exit status 2
+9 verbose stack     at EventEmitter.<anonymous> (/usr/lib/node_modules/npm/lib/utils/lifecycle.js:214:16)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at EventEmitter.emit (events.js:172:7)
+9 verbose stack     at ChildProcess.<anonymous> (/usr/lib/node_modules/npm/lib/utils/spawn.js:24:14)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at ChildProcess.emit (events.js:172:7)
+9 verbose stack     at maybeClose (internal/child_process.js:827:16)
+9 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:211:5)
+10 verbose pkgid media-picker@1.0.0
+11 verbose cwd /home/die/Escritorio/media-picker
+12 error Linux 4.2.0-34-generic
+13 error argv "/usr/bin/nodejs" "/usr/bin/npm" "run" "build"
+14 error node v4.4.0
+15 error npm  v2.14.20
+16 error code ELIFECYCLE
+17 error media-picker@1.0.0 build: `browserify js/app.js > js/build.js -t [babelify --presets [es2015 react] ] && uglify -s js/build.js -o js/build.min.js && && watchify js/app.js -o js/build.min.js -v -t [ babelify --presets [ react ] ]`
+17 error Exit status 2
+18 error Failed at the media-picker@1.0.0 build script 'browserify js/app.js > js/build.js -t [babelify --presets [es2015 react] ] && uglify -s js/build.js -o js/build.min.js && && watchify js/app.js -o js/build.min.js -v -t [ babelify --presets [ react ] ]'.
+18 error This is most likely a problem with the media-picker package,
+18 error not with npm itself.
+18 error Tell the author that this fails on your system:
+18 error     browserify js/app.js > js/build.js -t [babelify --presets [es2015 react] ] && uglify -s js/build.js -o js/build.min.js && && watchify js/app.js -o js/build.min.js -v -t [ babelify --presets [ react ] ]
+18 error You can get information on how to open an issue for this project with:
+18 error     npm bugs media-picker
+18 error Or if that isn't available, you can get their info via:
+18 error
+18 error     npm owner ls media-picker
+18 error There is likely additional logging output above.
+19 verbose exit [ 1, true ]
